Convert data_vio_operation and most data_vio flags to 1 bit fields.
This clean-up should make the data_vio read and write paths a bit
easier to understand as it eliminates some subtle uses of the mapping
states of the new_mapped and mapped zoned_pbns. It is also a small but
useful prequel for the follow-on to VDOSTORY-271 which eliminates the
need to synthesize dec refs by doing the block map lookup before making
the increment entry in the journal. Without eliminating the subtle use
of the mapping states, the lookup eliminates the data_vio's knowledge
of whether it is a trim.

Pair: raeburn

.../data-vio.h
Replaced the io_operation field of data_vio with three 1 bit fields:
  read, write, and fua. While the bit fields don't eliminate the
  bitwise operations necessary to extract the operation, the compiler
  will do it for us, probably more efficiently, and the resulting code
  is clearer. Renamed is_zero_block field to is_zero and made it a 1
  bit field. Converted is_trim, is_partial, and is_duplicate to 1 bit
  fields. These 7 fields now consume a single byte, rather than 17
  bytes.
Removed the data_vio_operation_bits and data_vio_operation enums along
  with the #defines of DATA_VIO_READ_MODIFY_WRITE and
  DATA_VIO_READ_WRITE_MASK as the bit fields for read, write, and fua
  can now be accessed directly.
Removed predicates which are now superceded by the bit fields:
  is_read_data_vio(),
  is_write_data_vio(),
  is_read_modify_write_data_vio(),
  is_trim_data_vio(),
  and data_vio_requires_fua().

.../data-vio.c
Modified launch_bio() and finish_cleanup() to set the data_vio's new
  bit fields directly, rather than passing an operation to
  launch_data_vio().
Removed the operation parameter from launch_data_vio() as the bit
  fields which have replaced the data_vio io_operation field are
  already set correctly by both of this functions callers.
Removed the setting of mapping states from launch_data_vio(). The
  new_mapped state was done as a way of differentiating trims from
  normal writes, but has been superceded by the new is_trim flag which
  is less subtle. The mapped state is now set when needed.
Modified continue_data_vio_with_block_map_slot() to set the new_mapped
  mapping state for zero block writes rather than doing in
  launch_data_vio() as this is where that setting matters.

.../block-map-tree.c
.../block-map.c
.../compression-state.c
.../data-vio.c
.../dump.c
.../recovery-journal-block.c
.../blockMapUtils.c
.../journalWritingUtils.c
Reflected changes.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157219]