Fixed bug where a partial discard discards the entire block [VDO-5489].
Change 157219 attempted to eliminate the subtlety of using a data_vio's
newmapped mapping state to determine whether or not a write requires an
allocation. Unfortunately, the existing logic was too subtle, and the
attempt was incorrect.

Pair: raeburn, sclafani

.../data-vio.c
Modified continue_data_vio_with_block_map_slot() to always allocate a
  block for partial discards when the rest of the block is not all
  zero.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157223]