Converted top-level check-style rule to use the kpatch directory.
The kpatch directory's prepare step fails to work without a defined
kernel version or an existing kernel tree. As of this change, when
running check-style from the top level, if no kernel version is
specified, the version of the running kernel (from uname -r) will be
used.

Pair: msakai

.../defines
Added PACKAGING_DIR and KPATCH_DIR.
Added setting of CHECK_PATCH_KERNEL_VERSION from KERNEL_VERSION if it
  is defined, or from uname -r if not.

.../src/Makefile
Modified check-style target to build in the kpatch directory rather
  than the github directory and to supply a kernel version.
Removed check-checkincludes: and check-checkpatch: targets as these are
  meaningless in the github directory and not defined in the kpatch
  Makefile.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157228]