VDO-5493 Run nfsGenData tests on suitable machines.
The NFS genData tests were moved to local hardware with the
NFS server on a bare metal machine many years ago due to
performance and network bandwidth limitatations of the cloud
resources available at the time. With the current test
resources, it is more efficient to run client and server on
VMs on the same subnet. This change also allows re-enabling
NfsGen03, which was excluded from the NfsGenData suite in
response to VDO-2249.

Pair: msakai, sclafani

.../GenData03.pm
Use the default nfsclientClass.

.../vdotests.suites
Don't exclude GenData03 from the nfsGenData suite.
Move the nfsGenData suite from the local test suite to the
  default suite but exclude from beaker testing.
Make XFS on the default client class the filesystem for NFS
  tests instead of ext3 on a bare metal machine.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157286]