[VDOUP-6] Add kstrtox.h with kstrtouint() addressing checkpatch issue.
checkpatch produces a warning when using sscanf() with a single
character format specifier to use kstrto<x> in its place.  This
changeset adds a vdo/fake kstrtox.h providing for non-kernel
compilation.

The implementation of kstrtouint() provided as part of vdo/fake is
based on strtoll() and mimics the underlying kernel implementation.
The mimicry is not perfect in that a more restricted range of values
are accepted as valid than the kernel would accept.  This
implementation explicitly disallows bases greater than 16 where as
the kernel's implementation (in v6.0.7) indirectly supports bases
beyond 16 (in contradiction of documentation) as long as the individual
digits come from the hexadecimal set.

Pair: msakai, jwiele

.../kstrtox.h
New file; provides vdo/fake kstrtouint() for non-kernel compilation.

.../pool-sysfs.c
Use kstrouint for single variable sscanf.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157287]