[UDSUP-1] Address checkpatch PRINTK_WITHOUT_KERN_LEVEL issue.
Although the appropriate kernel level value is included it is
included as an argument to a format string to printk().  This is
something that checkpatch does not recognize and so reports the
PRINTK_WITHOUT_KERN_LEVEL issue.

This changeset rewrites logging in such a way as to include the
appropriate kernel level in calls to printk().  This, unfortunately,
results in checkpatch reporting an issue (PREFER_PR_LEVEL) about using,
for example, pr_crit(...) rather than printk(KERN_CRIT).  Worsening the
issue is that the kernel does not provide a pr_default() method
irrespective of checkpatch's implication that it does.

So, to support detection of future failures to provide a kern level to
printk() this changeset addresses the PRINTK_WITHOUT_KERN_LEVEL issue
and suppresses the PREFER_PR_LEVEL issue in order to specify
KERN_DEFAULT as appropriate.

Pair: joe

.../logger.c
Rewrote logging to resolve to invocations of printk() with specified
  kernel levels.

.../Makefile
Added PRINTK_WITHOUT_KERN_LEVEL to list of checkpatch regression types.
Added PREFER_PR_LEVEL to list of suppressed checkpatch issues.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157294]