Renamed structs, fields, and functions in compression-state.[hc].
This is another prequel to combining compression-state with data-vio
for [VDOUP-6]. In preparing for that merger, it seemed that the naming
in compression-state could be improved. For example, having struct
compression_state as a sub-structure of data_vio seems odd when struct
data_vio_compression_state (the unpacked version of the
compression_state's atomic) gets pulled in next to it.

Pair: sclafani

.../compression-state.h
.../compression-state.c
Renamed vio_compression_status enum to data_vio_compression_stage, and
  renamed all of the enumerants to be DATA_VIO_* rather than VIO_*.
  Renamed all functions and fields which used to be *state* to
  *status*.
Renamed struct vio_compression_state to data_vio_compression_status,
  and renamed the status field to stage.
Renamed all _vio_ functions to be _data_vio_ since they only operated
  on data_vios.

.../data-vio.h
Renamed the state field of of compression_state to status.

.../data-vio.c
.../dedupe.c
.../packer.c
.../Moot_t1.c
Reflected changes.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157442]