Fixed problems loading new tables [VDO-5508].
Change 157456 which attempted to fix VDO-5505 was still not quite
right. When a new table is loaded for an existing vdo, the new
device_config should not be installed in the vdo until pre-resume,
rather than doing table load.

Pair: raeburn

.../dm-vdo-target.c
Removed the setting of the vdo's device_config field from
  set_device_config(), adding it to just after the call to
  set_device_config() in initialize_vdo(). By so doing, the field is
  set for new vdos, but not modified ones.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157472]