Declared a function noinline to avoid valgrind failures [VDO-5506]
This is the same issue as VDO-5182, but in a new function. See
change 149367 for a detailed explanation, but to sum up: valgrind
is complaining about uninitialized data in a structure that is
clearly initialized in the code. This might be a GCC bug or just a
spurious valgrind error, but to be safe we can declare the
function noinline to avoid triggering the problem.

Pair: corwin

.../vdo-recovery.c
Declared count_increment_entries() noinline and added an XXX comment
  summarizing why.	

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157474]