Eliminated abort_data_vio_optimization().
This function is really the end of the compression path for data_vios
which didn't get compressed. It was pointed out that its name is
somewhat confusing since it is called from the dedupe path on hash lock
agents which have waiters and need to write their blocks. Technically,
optimization for the agent isn't aborted (in one sense), since its
waiters will deduplicate against it.

Pair: sclafani

.../data-vio.h
.../data-vio.c
Pulled abort_data_vio_optimization() into write_block(), renamed it to
  write_data_vio(), and exposed it.

.../data-vio.c
.../dedupe.c
.../packer.c
Replaced all calls to abort_data_vio_optimization() with calls to
  write_data_vio().

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157479]