Fixed infinite loop in hash lock continuation [VDO-5483].
If the lock agent gets an error, it will put the hash lock into the
BYPASSING state when it gets to the hash lock cleanup stage. If there
are other data_vios which managed to successfully perform their metadata
updates before the vdo went read-only due to the agent's error, they
will observe this state when they return to give up their share of the
lock.

Unfortunately, continue_hash_lock() makes the incorrect assumption that
any data_vio whose lock is already BYPASSING must have an error since
the lock agent must have had an error which would have put the vdo into
read-only mode. In this case, we call continue_data_vio(). When the
data_vio doesn't have an error, this doesn't invoke the error handler,
as expected, it merely invokes the callback, which is still set to
continue_hash_lock().

Pair: sclafani, raeburn

.../dedupe.c
Modified continue_hash_lock() to call exit_hash_lock() if the lock is
  BYPASSING.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157491]