Fix problem when compiled with clang.
We have two problems with compiling under clang:
1. ./uds/logger.c:309:49: error: format string is not a string literal 
           [-Werror,-Wformat-nonliteral]
           uds_log_embedded_message(priority, NULL, NULL, format, args, "%s", "");

2. kvdo/work-queue.c:600:32: error: unknown warning group '-Wformat=', 
   ignored [-Werror,-Wunknown-warning-option]
          #pragma GCC diagnostic ignored "-Wformat="

To fix the first error, we need to remove -Wformat-nonliteral which is
turned on by -Wformat=2 in defines file.  Therefore, we move -Wformat=2 
in gcc session only.

To fix the second error, we remove the '=' sign in the #pragam, which
should archive the same goal. 

Pair: jshimkus

.../defines
Move -Wformat=2 to the gcc warning session.

.../work-queue.c
Remove the '=' sign which cause problem with clang.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157562]