Fix kpatch/Makefile to allow it to patch branches.
Pair: jshimkus

.../Makefile
Remove the drivers/md/dm-vdo directory before repopulating it
  so that files being deleted can't stick around.
Update the drivers/md/Makefile and drivers/md/Kconfig relative
  to the master branch so that deleted lines can't stick
  around.
Clean the current branch (which could be the master), rather
  than always resetting to the state of the master branch.
Generate the patch relative to the current branch (which could
  be the master), rather than the unconditionally generating
  the patch against the master branch.
Add checkincludes.out to the list of files to clean.
Update comments.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157592]