Fix id-exclusions target.
The id-exclusions target is used to exclude directories when building
the TAGS file. Unfortunately, with the removal of src/packaging/rpm,
the invocation "test -d packaging/rpm &&" causes the rule to fail.

Pair: raeburn

.../Makefile
Replaced test -d foo && with if [ -d foo ]; in the id-exclusions target
  as the latter does not cause the target to fail if the directory does
  not exist.
Added exclusion of the entire packaging directory to id-exclusions so
  that prepared directories are not examined.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157669]