Fixed rare hang in JournalThresholds_t1 [VDO-5432].
This test attempts to set certain conditions on the recovery journal,
and then attempts to validate when the journal is not
reaping. Unfortunately, the method it used to check for not reaping can
deliver a false positive as the journal may be about to reap when the
check is made. This change switches to a more reliable mechanism.

The test initialization attempts to get the vdo into a state where the
physical space has been filled, but the recovery journal is about to
start a new block and has been fully reaped. Rather than doing this by
carefully writing just enough data to fill the last dirty journal
block, and then forcing a reap, it is simpler to merely restart the VDO
after filling. This also eliminates the need to wait for reaping to
complete before starting the test.

Pair: sclafani

.../JournalThresholds_t1.c
Rewrote initializeTest() to fill and then restart the VDO, and only
  thereafter adjusting the size of the recovery and slab journals. Also
  eliminating caching of the slab depot and moved the initialization of
  the slabSummaryWriteCount state variable to just before it is used in
  keeping with our current idiom.
Renamed interrogateJournal() to sampleJournal() and called it directly
  rather than through interrogateJournalUntilNotReaping() in places
  where we only want to sample the journal.
Fixed races in testRecoveryJournalThreshold():
  - Set a new hook, failOnSlabJournalReap() which will cause the test
    to fail if the journal attempts to reap the slab journal head
    prematurely. The spelling of this hook necessitates switching this
    test to not use the 1-thread config (which also makes it a more
    realistic test).
  - Replaced the failure hook with a new hook to signal when the
    journal has reaped and wait for that state in addition to the
    expected slab summary writes before calling
    interrogateJournalUntilNotReaping(). This ensures that the reap we
    expected has started before we check to see if it is done.
Respelled testRecoveryJournalThreshold() to not hard code recovery
  journal sequence numbers, but rather base them on state of the
  journal when the test starts. This makes the test easier to modify,
  particularly for the upcoming removal of synthesized decrefs.
  Updated testScrubSlabDuringRebuild() to reflect the new initial state
  of the journal.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157705]