Fixed rare hang in BlockMapTreeWrites_t1 [VDO-5008].
This test (specifically testBlockMapTreeGenerationRollOver()) attempts
to carefully and precisely advance the recovery journal past the block
map threshold in order to cause tree pages to be written out. However,
when multiple data vios are writting concurrently, it is possible for
the journal to attempt an era advance while the block map's action
manager is still performing the previous era advance. When this
happens, the latter advance is ignored (under the assumption that we
have plenty of head room, and missing a few notifications is not a big
deal). However, with the precise write pattern of this test, the missed
era advance results in a hang.

Pair: sclafani

.../BlockMapTreeWrites_t1.c
Modified testBlockMapTreeGenerationRollOver() to break up the batch of
  writes which attempt to advance the journal past the maxium age and
  to wait for the block map's action manager to not be operating
  between batches. This ensures that no era advances are missed.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157706]