Fixed valgrind complaint in fake submit_bio_wait() [VDO-5516]
The unit test implementation of submit_bio_wait() creates a vio on the
stack if the submitted bio doesn't have one. This is necessary for that
bio to be processed as all others by async layer. While this is
currently not strictly necessary, given the limited uses of
submit_bio_wait(), it does correctly maintain the invariant that all
accesses to the RAMLayer are done on the bio processing thread.

Pair: sclafani

.../testBIO.c
Modified submit_bio_wait() to minimally initialize the vio it creates
  on the stack when the submitted bio doesn't have one. Also added an
  assertion that any bio submitted without a vio has a vcnt of 0 since
  this ensures that processBIO() will not attempt to access any fields
  of the vio other than its completion's parent.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157709]