Made vdo_page_cache more block map specific [VDOUP-6].
The vdo_page_cache was implemented to be a generic cache, not block map
specific. However, it is only used by the block map, and it is likely
that if vdo ever needs another cache, it will either be able to use an
existing kernel facility, or will need different semantics. This change
is a small step towards making the page cache fully aware of the block
map pages it caches. It is also a prequel to pulling vdo_page_cache.c
into block_map.c.

Pair: msakai

.../vdo-page-cache.h
.../vdo-page-cache.c
.../block-map.c
Pulled the read hook and write hook implementations in block-map.c into
  their callers in vdo-page-cache.c. Since the page cache is only used
  by the block map, these hooks need not be generic. Eliminated the
  read_hook and write_hook fields of struct vdo_page_cache, their types
  (vdo_page_read_function and vdo_page_write_function), and constructor
  parameters:
  - Merged run_read_hook(), page_is_loaded(), and validate_page_on_read().
  - Merged page_is_written_out() and handle_page_write().
Replaced the context field of struct page_info with a recovery_lock
  field as this is what the generic context is always used
  for. Eliminated the page_context_size parameter from
  vdo_make_page_cache(). Removed vdo_get_completion_context() as the
  recovery_lock is now accessible directly from the page_info.
Exposed as_vdo_page_completion().

.../VDOPageCache_t1.c
Rewrote this test to reflect changes:
- Switched from a basic test to a vdo test as the page cache now
  requires a recovery journal. Also added an initialization step to
  increment the lock counts on all of the recovery journal blocks for
  which page writes will attempt to release locks. Adapted to the fact
  that in a real vdo, the page number of a block map page can not be
  the same as its pbn.
- Replaced the read and write hooks with a callback enqueue hook to
  wrap the callback from a page I/O's endio callback.
- Eliminated the custom page format, using real block map pages instead
  and relying on the validation in the page cache itself for some of
  the checks.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157728]