Rearrange the start of recovery [VDOSTORY-271].
This change reworks the start of recovery. The primary aim of this
restructuring is to pave the way for the elimination of synthesized
decrefs, which needs to introduce a new error path so that attempting
to recover an old format journal results in the vdo going read only.

Pair: msakai

.../vdo-recovery.c
Pulled find_contiguous_range() into prepare_to_apply_journal_entries(),
  and respelled this method to continue the completion in all cases
  where there are valid recovery journal entries or where there is an
  error. Moved the handling of successful cases where there are no
  journal entries from prepare_to_apply_journal_entries() up to
  launch_recovery().
Factored checking that the journal heads make sense relative to the
  tail into new validate_heads() since it is now called both from
  prepare_to_apply_journal_entries() and launch_recovery().
Modified launch_recovery() and launch_rebuild() to handle their own
  errors and not return an error code to finish_journal_load() since
  the error handling is going to get more complicated, and the
  unfactoring of the error handling is clearer as finish_journal_load()
  no longer needs to decide whether or not to finish the parent.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157745]