Added entry count validation of journal sectors [VDOSTORY-271].
This is another prequel to the elimination of synthesized
decrefs. Because we will need to suport both the old and new recovery
journal formats, and the number of entries per sector is different in
each, it seems prudent to validate that the sector entry counts are not
too large.

Pair: sclafani

.../vdo-component-states.h
Modified vdo_is_valid_recovery_journal_sector() to take the sector
  number and check that the entry count is not too high. The sector
  number is required since the last sector has a lower maximum count
  than the others.

.../vdo-recovery.c
.../vdoDebugMetadata.c
Reflected changes.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157757]