Created a (disabled) test of recovery journal upgrade [VDOSTORY-271].
The elimination of synthesized decrefs will result in a new recovery
journal format. For upgrades of clean vdos, this is no issue since the
journal will be ignored. However, an upgrade of a dirty vdo can't do a
normal recovery unless we continue to maintain the ability to
synthesize decrefs. Fortunately, read-only rebuild does not synthesize
decrefs, so the final implementation will go into read-only mode if
recovery detects that the journal is in the old format. While we will
need to preserve the old journal format for rebuild, this is far less
code and complication than that for synthesizing decrefs.

For now, the test is disabled as the dirty test case does not pass
since recovery of the old format is still supported.

Pair: sclafani

.../OldRecoveryJournal_t1.c
.../vdo.old.rj.crashed
.../vdo.old.rj.recovered
.../Makefile
Created new (disabled) unit test and the two pickles it requires.

.../testParameters.h
.../testParameters.c
Added a 'backingFile' field to TestParameters and TestConfiguration so
  that tests can specify a pickle to back the RAMLayer at
  initialization time.

.../vdoTestBase.h
.../vdoTestBase.c
Exposed initializeTest() as the new test needs finer-grained control of
  initialization than was available through the higher level
  initializers.
Modified initializeTest() to use the backingFile if one has been specified.
Modified initializeVDOTest() to not format the vdo if a backingFile was
  specified, otherwise the backingFile has no effect.

.../JournalReplay_t1.c
.../JournalThresholds_t1.c
.../RollOver_t2.c
.../SynthesizeDecRef_t1.c
Renamed the initializer to not conflict with the newly exposed
  initializeTest() from vdoTestBase.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157759]