Clean up some comments, removing XXX.
Pair: corwin

.../dedupe.c
Clarify the purpose of the BYPASSING state slightly and remove
  the XXX note about it.
Tighten the note about the correlation of function names to
  states without the reference to VDOSTORY-190. I think the
  gist is clear enough with the example, and in fact may be
  unnecessary.
Delete an XXX comment describing a line of code updating the
  max_references counter that seems perfectly clear on its
  own without a comment.
Change some "XXX" to "TODO:".
Correct spelling of "minuscule".
Delete comments about overlapping fragments of the record name
  hash as all 128 bits are used by UDS. The two bit fields
  extracted from the hash in this file do not overlap, but
  does that even matter since they are used for different
  purposes?

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157786]