Fixed VDO-5523 ValgrindUnitTests::OldRecoveryJournal_t1 failure
The fix for VDO-5523 in change 157792 missed the code paths where
the filename is passed via test parameters and so was still using
"testdata/<file>" constants as paths.

Pair: slegendr

.../OldRecoveryJournal_t1.c
Added globals to hold the full pathnames.
Moved the pathname construction to initialize().
Factored a cleanUp() method to free the pathnames.
Passed the generated pathnames as test parameters.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157829]