Eliminate dirty_lists.[hc] [VDOUP-6].
This change moves the dirty_lists implementation into block-map.c. It
also leverages the fact that both dirty lists clients (the page cache
and the tree pages) are in one file and more aware of each other to
allow each block map zone to have a single dirty_lists structure which
serves both its cache and tree. Currently, each era in the dirty lists
has a separate list for tree pages and cache pages, but in future, it
may be possible to find an inexpensive way to share the lists.

Pair: msakai

.../dirty-lists.h
.../dirty-lists.c
.../block-map.h
.../block-map.c
Moved the dirty_lists implementation into block map:
- Defined dirty_era_t to be an array of two lists heads, one for cache
  pages, and one for tree pages. Defined block_map_page_type enum to
  index into a dirty_era_t. This enables each block map zone to have
  only a single set of dirty lists.
- Pulled write_dirty_cache_pages_callback() and
  write_dirty_pages_callback() into write_expired_elements(). Added the
  block map zone as a parameter to this function which obviates the
  need for the dirty_lists context.
- Made vdo_add_to_dirty_lists() EXTERNAL_STATIC (VDOPageCache_t1 calls
  it) and dropped the vdo_ prefix. Added a type parameter and modified
  the function to handle both tree and cache pages.
- Removed the callback and context fields from struct dirty_lists since
  the lists can now be aware of how to expire dirty elements directly.
- Pulled vdo_make_dirty_lists() into initialize_block_map_zone() and
  reflected the new structure handling both types of pages.
- Pulled vdo_advance_dirty_lists_period() into
  advance_block_map_zone_era().
- Pulled vdo_set_dirty_lists_current_period() into
  vdo_initialize_block_map_from_journal().
- Pulled vdo_flush_dirty_lists() into initiate_drain(). 
Moved the INTERNAL block of EXTERNAL_STATIC prototypes to the end of
  block-map.h as there is no better place to have a single such block.
Removed the dirty_lists field from struct vdo_page_cache and the
  construction of it from allocate_cache_components(). Also removed the
  maximum_age parameter from this function as it was only used to
  construct the dirty_lists. Removed the clean up of this field from
  uninitialize_block_map_zone().

.../Makefile.common
.../MANIFEST.yaml
Eliminated dirty_lists.[hc].

.../VDOPageCache_t1.c
Reflected changes.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157839]