Reduced chance of losing a race in Suspend_t1 [ALB-3111]
suspendRebuildTest() launches a thread and then tries to catch it
in the midst of rebuilding, which is fundamentally racy. This
change adjusts the duration of sleeps and the rebuild in an
attempt to avoid the occasional failure under valgrind due to
losing the race.

Pair: msakai

.../Suspend_t1.c
Increased number of chapters from 7 to 10 to try to decrease the
  chance of the test failing to see an in-progress rebuild.
Halved the sleep duration (20ms->10ms) to try to have more chances
  of seeing an in-progress rebuild.
Added polling/sleep loops instead of always sleeping a total of a
  half second waiting for rebuild to finish. This somewhat
  ameliorates the increased test run time due to increasing the
  number of chapters.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157856]