Restore the "slab count" stat [VDO-5524].
Pair: sclafani

The "slab count" stat was derived from the slab count of the
allocator until change 157838, which changed it to use the
slab count of the block allocator statistics. Unfortunately,
nothing writes the slab count in the statistics so the "slab
count" stat is always reported as 0.

The slab_count in the block allocator statistics should either
be made useful or removed.

.../slab-depot.c
Derive the "slab count" stat from the actual slab counts
  rather than the unused slab_count in the allocator
  statistics.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157857]