Merge change 152919 to linux-vdo.
Adjustments have been made for Sulfur and Chlorine versions and
scenarios instead of Aluminum ones.

The Upgrade device was modified to use an LVMManaged base class.

The LVMManaged device was modified to set parameters that need
to be defined prior to module installation.

Pair: msakai 

.../Permabit/BlockDevice/VDO/Upgrade.pm
.../UpgradeBase.pm
.../MigrationBase.pm
.../vdotest/VDOTest/Upgrade.pm
.../UpgradeCompressDedupe.pm
.../UpgradeLatest.pm
.../Downgrade.pm
.../DirtyUpgrade.pm
.../SupportedScenarios.yaml
.../SupportedVersions.pm
.../SupportedVersions_t1.pm
152919: Enhance upgrade test framework to use scenarios.

.../Permabit/BlockDevice/VDO/LVMManaged.pm
Add installModule() to set the module path if it is not set.
  This is needed to account for LVMManaged devices not going
  through BlockDevice::VDO::new().

.../SupportedVersions.yaml
Adjust the versions to align with the supported scenarios.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157938]