Switched the instance tracker to a spinlock [VDOUP-6]
Kernel style is to declare lock globals statically with an
initializer (via a macro) instead of running init functions. It
doesn't need to be a mutex as we don't do operations while holding
the lock that would require it.

Pair: jwiele

.../dm-vdo-target.c
Moved the instances lock out of the structure.
Converted it to a spinlock_t by declaring it with DEFINE_SPINLOCK().
  Removed its init and destroy calls.
Switched the lock/unlock calls to spin_lock() and spin_unlock(),

.../spinlock.h
Added DEFINE_SPINLOCK().

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157961]