Fixed vdoDebugMetadata [VDO-5529].
Change 157949 modified the decoding of the layout such that it relies
directly on the geometry. The way in which vdoDebugMetadata manipulates
the geometry to reflect the dump runs afoul of the new validation that
was added to layout decode.

Pair: msakai

.../vdoDebugMetadata.c
Modified readFromDump() to load and decode the super block and
  component states directly rather than using loadVDOWithGeometry() as
  loading the super block requires the modified geometry, but decoding
  the component states requires the original geometry.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157966]