Simplify use of struct buffer in UDS. [UDSUP-1]
Pair:  corwin

.../config.h
.../config.c
Made on-disk structure packed, and changed decode functions to
  unmarshall into in-memory structures instead.
Refactored structure encode and decode to share common code.

.../config.c
.../index-layout.c
.../volume-index.c
.../delta-index.c
.../index-page-map.c
Replaced uses of struct buffer with a simple character buffer,
  calling marshalling functions from numeric.h directly. The error
  checking provided by struct buffer is not useful the way UDS works,
  and skipping the interface removes a lot of unnecessary error
  checking.
Inlined simplified encode and decode functions.

.../delta-index.c
Removed a check for a saved zone number out of range because it is
  redundant with the subsequent check that the saved zone number is
  the one we expected. (The index never expects an out-of-range zone
  number.)

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157986]