Updated vdodumpconfig tool manpage [VDO-5528] [BZ 2176778]
Pair: msakai

.../vdodumpconfig.8
Removed claim that the tool works on a running VDO since LVM will not
  allow access to the backing store while VDO is running.
Noted that the backing store needs to be a VDODataLV as described
  on the lvmvdo man page.
Added an example of using lvchange to activate and deactivate the
  VDODataLV when invoking the tool.
Linked to the lvmvdo and lvchange manpages.
Removed link to vdo (manager) man page.

.../vdoDumpConfig.c
Updated usage string correspondingly.
Fixed cut&paste error in usage string.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158104]