Remove impossible assertions from volume.
Pair: jwiele

.../volume.h
.../volume.c
Removed the assertion in get_least_recent_page() that we have a page
  with the read_pending flag unset. Any given thread can only set the
  read_pending flag on one page, so there can be at most 17 pages
  with this flag set (16 read threads plus the chapter writer
  thread). 17 is smaller than any possible cache size. (The rebuild
  process can also set this flag, but during rebuild the other threads
  are not active.)
Changed get_least_recent_page() to return the page now that it does
  not return errors.
Removed the assertion in select_victim_in_cache(). This assertion was
  redundant with the assertion in get_least_recent_page(), and by the
  same logic that function can never return a null pointer.
Changed select_victim_in_cache() to return the selected page now that
  it does not return errors.
Removed the assertion in put_page_in_cache() that the given page is
  in the proper range. The earlier assertion in this function
  guarantees that the page came from select_victim_in_cache(), which
  cannot return a page that is not in the cache. Also inlined value
  into its only remaining use.
Removed the assertion in enqueue_read() that the read_queue_index is
  in the proper range. The logic of setting and advancing the
  read_queue_index immediately prior guarantees this fact.
Changed the ASSERT_LOG_ONLY plus error return at the end of
  init_chapter_index_page() to use the more standard idiom of
  uds_log_error_strerror().
Removed the assertion in search_cached_index_page() that a search is
  pending. This is trivally true because this function has already
  called begin_pending_search().
Removed the assertion in find_chapter_limits() that left_chapter and
  right_chapter are the same. This is always true as the preceding
  algorithm guarantes that left_chapter is always less than or equal
  to right_chapter. Also respelled the subsequent mod calculation
  to avoid the mod and explain more clearly what it's doing.
Removed the assertion in find_chapter_limits() that the lowest chapter
  number is valid, since the preceding algorithm will have already
  determined that.

.../PageCache_t1.c
.../PageCache_p1.c
Reflected changes.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158129]