Add some additional rebuild testcases for converted indexes.
Recently, I thought I found a bug around finding the volume boundaries
during index rebuild, if conversion had moved a chapter the very end
of the physical volume. Fortunately, I was wrong and there is no bug.
But it seemed prudent to add some testcases for the scenario I was
concerned about, since it was not previously tested explicitly.

Pair: sclafani

.../Volume_t3.c
Added some testcases for conversion moving a chapter to the
  of the physical volume.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158168]