Update SRPM build for linux-vdo.
Pair: slegendr

.../src-dist/MANIFEST.yaml
Install kvdo sources in $(src)/dm-vdo/. dm-vdo-target.c will be
  moved to $(src) by the Makefiles.
Do not bother to undefine __KERNEL__ for sources that are
  userspace only.
Include dory and histogram sources; they are needed by the
  SRPM build, but will be filtered out of the github and
  kpatch builds.

.../src-dist/kernel/Makefile
Build a single dm-vdo module, using a wildcard to locate the
  object files.

.../src-dist/kvdo.spec
Build a single dm-vdo module.
Remove stuff that is obsolete and/or irrelevant to a DKMS
  build.
Update comments.

.../rpm/postProcess.sh
New script to replace EXTERNAL_STATIC with the empty string,
  analogous to the .../github/removeInternal.sh script that
  replaces "EXTERNAL_STATIC" with "static".

.../rpm/MANIFEST.yaml
Exclude histogram sources from the user utilities build.
Use the .../github/removeInternal.sh script to replace
  "EXTERNAL_STATIC" with "static" for the user utilities
  build.
Use the new postProcess.sh script to replace "EXTERNAL_STATIC"
  with "" for the kernel module build so test functions can be
  found.
Restrict the undefines to just VDO_USER for the kernel module
  SRPM so that test code gets built in.

.../rpm/Makefile
Use a two step process to generate the SRPM, similar to the
  process used in the kpatch directory: first prepare the
  source files with prepareSourceDist.pl, then move
  dm-vdo-target.c to $(src), and finally recreate the tarball.
Use a callable macro instead of recursive make to make each
  SRPM.

.../github/MANIFEST.yaml
Exclude test only files dory.[hc] and histogram.[hc] from the
  builds.
Include the COPYING and Makefile files that are only relevant
  to the github push.

.../github/Makefile
Move dm-vdo-target.c from $(src)/dm-vdo to $(src).

  .../kpatch/MANIFEST.yaml
Exclude test only files dory.[hc] and histogram.[hc] from the
  kernel module build.

.../kpatch/Makefile
Adjust to the change in the src-dist/MANIFEST.yaml that puts
  the sources in the dm-vdo directory.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158175]